+ Atualizar modelagem do banco de dados + Padronizar nome dos atributos das entidades ao padrão do spring data + Criar entidade para representar a operadora do cartão de crédito

+++ src/main/java/com/mymoney/repositories/CategoryRepository.java
@@ -11,4 +11,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Category, Short> { }
+public interface CategoryRepository extends JpaRepository<Category, Long> { }
